requirements.txt
tensorflow-cpu
gradio